src/MCPServer/requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=dev.txt dev.in
#
atomicwrites==1.3.0
attrs==19.1.0
click==7.0                # via pip-tools
django-extensions==1.1.1
django==1.8.19
enum34==1.1.6
filelock==3.0.10
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
gearman==2.0.2
grpcio==1.22.0
jsonschema==2.6.0
lxml==3.5.0
mock==2.0.0
more-itertools==5.0.0
mysqlclient==1.3.7
pathlib2==2.3.3
pbr==5.1.3
pip-tools==3.7.0
pluggy==0.9.0
prometheus_client==0.7.1
protobuf==3.9.0
py==1.8.0
pytest-django==3.4.8
pytest-mock==1.10.2
pytest-pythonpath==0.7.3
pytest==4.3.1
scandir==1.10.0
six==1.12.0
toml==0.10.0
tox==3.8.3
virtualenv==16.4.3

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via protobuf, pytest, tox
